add drop database exception throwing

+++ src/data/DatabaseManager.java
@@ -1,9 +1,10 @@
 package data;
 
-import data.Exceptions.DatabaseNotFoundException;
-
 import java.io.File;
+import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Created by khaled on 11/26/17.
@@ -13,9 +14,11 @@ public class DatabaseManager implements IDatabaseManager {
     private static DatabaseManager databaseManager;
     public ArrayList<String> allDatabase = new ArrayList<>();
     private static final String databasesPath = "Database" + System.getProperty("file.separator");
+    private Map<String, String> databasesPathMap;
 
     private DatabaseManager() {
         databaseInUse = null;
+        databasesPathMap = new HashMap<>();
     }
 
     public static DatabaseManager getInstance() {
@@ -34,8 +37,9 @@ public class DatabaseManager implements IDatabaseManager {
     }
 
     @Override
-    public void dropDatabase(String databaseName) throws DatabaseNotFoundException {
+    public void dropDatabase(String databaseName) throws SQLException {
         File databaseDir = new File(databasePath(databaseName));
+        if (!databaseDir.exists()) throw new SQLException();
         deleteFilesInDirectory(databaseDir);
         databaseDir.delete();
         databaseInUse = null;
@@ -47,7 +51,7 @@ public class DatabaseManager implements IDatabaseManager {
     }
 
     @Override
-    public void setDatabaseInUse(String databaseName) throws DatabaseNotFoundException {
+    public void setDatabaseInUse(String databaseName) throws SQLException {
         databaseInUse = new Database(databaseName);
     }
